requirements.txt
pyrtlsdr
scipy
numpy
pyaudio
